requirements.txt
openai==0.28.1
langchain==0.0.336
python-dotenv
